requirements.txt
faiss-cpu
numpy
ollama
openai
pypdf
sentence-transformers
streamlit
